Moved \Bolt\BoltUrlMatcher to \Bolt\Routing\UrlMatcher. Created a RoutingServiceProvider.
