+++ src/Routing/UrlMatcher.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bolt;
+namespace Bolt\Routing;
 
 use Silex;
 use Symfony\Component\Routing\Exception\ResourceNotFoundException;
@@ -11,7 +11,7 @@ use Symfony\Component\Routing\Exception\ResourceNotFoundException;
  *
  * This additionally checks for a route without a trailing slash and redirects to it.
  */
-class BoltUrlMatcher extends Silex\RedirectableUrlMatcher
+class UrlMatcher extends Silex\RedirectableUrlMatcher
 {
 
     public function match($pathinfo)
